Commands: Combine handler source files
Resolves #1771
