+++ test/EntityFramework.Commands.Tests/project.json
@@ -1,9 +1,12 @@
 {
+  "compilationOptions": {
+    "define": [ "OMIT_HANDLER_INTERFACES", "ENABLE_HANDLERS" ]
+  },
   "dependencies": {
     "EntityFramework.Commands": "7.0.0-*",
     "EntityFramework.Relational.Tests": "1.0.0"
   },
-  "compile": "..\\..\\src\\EntityFramework.Commands\\tools\\Handlers.cs",
+  "compile": "..\\..\\src\\EntityFramework.Commands\\Handlers.cs",
   "commands": {
     "test": "xunit.runner.dnx"
   },
